requirements.txt
streamlit==1.26.0
Pillow==9.5.0
pytesseract==0.3.10
nltk==3.8.1
sumy==0.10.0
scikit-learn==1.3.0
pyttsx3==2.90